requirements.txt
babel==2.17.0
pillow==11.1.0
python-dateutil==2.9.0.post0
six==1.17.0
tkcalendar==1.6.1
